requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
aiosqlite==0.20.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
Pillow==10.3.0
python-magic==0.4.27
aiofiles==23.2.1
httpx==0.27.0
pydantic-settings==2.2.1
openai==1.30.1
boto3==1.34.101
python-dotenv==1.0.1
